improve DELETE rest endpoint

+++ src/main/scala/omnibus/service/OmnibusService.scala
@@ -23,7 +23,7 @@ class OmnibusService(topicRepo: ActorRef, subscriberRepo: ActorRef) extends Acto
 
   def receive = {
     case CreateTopic(topic)                              => sender ! createTopic(topic)
-    case DeleteTopic(topic)                              => sender ! deleteTopic(topic)
+    case DeleteTopic(topic)                              => deleteTopic(topic) pipeTo sender
     case CheckTopic(topic)                               => checkTopic(topic, sender)
     case PublishToTopic(topic, message)                  => publishToTopic(topic, message) pipeTo sender
     case SubToTopic(topic, responder, reactiveCmd, http) => subToTopic(topic, responder, reactiveCmd, http) pipeTo sender
@@ -38,9 +38,23 @@ class OmnibusService(topicRepo: ActorRef, subscriberRepo: ActorRef) extends Acto
 
   def createTopic(topic: String) = topicRepo ! TopicRepositoryProtocol.CreateTopicActor(topic)
 
-  def deleteTopic(topic: String): String = {
-    topicRepo ! TopicRepositoryProtocol.DeleteTopicActor(topic)
-    s"Topic $topic deleted \n"
+  def deleteTopic(topic: String): Future[Boolean] = {
+    log.debug(s"Deleting topic $topic")
+    val p = promise[Boolean]
+    val f = p.future
+    val exist = (topicRepo ? TopicRepositoryProtocol.CheckTopicActor(topic)).mapTo[Boolean]
+    exist.onComplete {
+      case Failure(error) => p.failure { new Exception(s"an error occured while deleting $topic \n") }
+      case Success(topicExists) => {
+        if (topicExists) {
+          topicRepo ! TopicRepositoryProtocol.DeleteTopicActor(topic)
+          p.success(true)
+        } else {
+          p.failure { new TopicNotFoundException(topic)}
+        }
+      }
+    }  
+    f    
   }
 
   def checkTopic(topic: String, replyTo: ActorRef) = {
